added owner_id

+++ app/routers/body_measurement.py
@@ -5,7 +5,7 @@ from ..oauth2 import get_current_user
 from  .. import schemas,models
 from typing import List
 import math
-
+from sqlalchemy import asc,desc
 
 router = APIRouter(
     prefix='/bodymeasurements',
@@ -13,11 +13,27 @@ router = APIRouter(
 )
 
 @router.get('/',status_code=status.HTTP_200_OK,response_model=schemas.PaginatedBodyMeasuresResponse)
-def get_measurements(db:Session=Depends(get_db),current_user=Depends(get_current_user),limit:int=Query(10,ge=1,le=10),page:int=Query(1,ge=1)):
+def get_measurements(db:Session=Depends(get_db),current_user=Depends(get_current_user),limit:int=Query(10,ge=1,le=10),page:int=Query(1,ge=1),sort_by:str=Query('id'),order:str=Query('asc')):
+    
+    sort_fields = {
+        'id':models.BodyMeasurement.id,
+        'weight':models.BodyMeasurement.weight,
+        'chest':models.BodyMeasurement.chest
+    }
+
     total = db.query(models.BodyMeasurement).count()
     total_pages = math.ceil(total/limit)
     offset = (page-1) * limit
-    measurements = db.query(models.BodyMeasurement).limit(limit).offset(offset).all()
+
+    sort_column = sort_fields.get(sort_by,models.BodyMeasurement.id)
+    query = db.query(models.BodyMeasurement).filter(models.BodyMeasurement.owner_id == current_user.id)
+
+    if order == 'desc':
+        query = query.order_by(desc(sort_column))
+    else:
+        query = query.order_by(asc(sort_column))
+
+    measurements = query.limit(limit).offset(offset).all()
     return {
         'data':measurements,
         'total':total,
@@ -29,7 +45,7 @@ def get_measurements(db:Session=Depends(get_db),current_user=Depends(get_current
 
 @router.get('/{id}',status_code=status.HTTP_200_OK,response_model=schemas.BodyMeasurementResponse)
 def get_measurement(id:int,db:Session=Depends(get_db),current_user=Depends(get_current_user)):
-    db_measurement = db.query(models.BodyMeasurement).filter(models.BodyMeasurement.id == id).first()
+    db_measurement = db.query(models.BodyMeasurement).filter(models.BodyMeasurement.id == id,models.BodyMeasurement.owner_id == current_user.id).first()
     if db_measurement is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f'body measurement with id {id} not found')
     return db_measurement
@@ -38,7 +54,7 @@ def get_measurement(id:int,db:Session=Depends(get_db),current_user=Depends(get_c
 
 @router.post('/',status_code=status.HTTP_201_CREATED,response_model=schemas.BodyMeasurementResponse)
 def create_measurement(body_measurement:schemas.BodyMeasurementBase,db:Session=Depends(get_db),current_user=Depends(get_current_user)):
-    measurement_db = models.BodyMeasurement(**body_measurement.dict())
+    measurement_db = models.BodyMeasurement(**body_measurement.dict(),owner_id=current_user.id)
     db.add(measurement_db)
     db.commit()
     db.refresh(measurement_db)
@@ -52,6 +68,8 @@ def update_measurement(id:int,measurements:schemas.BodyMeasurementUpdate,db:Sess
     existing_db = db_measurement.first()
     if existing_db is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f'body measuremt with id {id} not found')
+    if existing_db.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail='Not authorized to perform the requested action')
     db_measurement.update(measurements.dict(exclude_unset=True),synchronize_session=False)
     db.commit()
     updated_measurement = db_measurement.first()
@@ -65,6 +83,8 @@ def update_measurement(id:int,body_measurement:schemas.BodyMeasurementBase,db:Se
     existing_db = db_measurement.first()
     if existing_db is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f'body measurement with id {id} not found')
+    if existing_db.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail='Not authorized to perform the requested action')
     db_measurement.update(body_measurement.dict(),synchronize_session=False)
     db.commit()
     updated_db = db_measurement.first()
@@ -76,6 +96,8 @@ def delete_measurement(id:int,db:Session=Depends(get_db),current_user=Depends(ge
     db_measurement = db.query(models.BodyMeasurement).filter(models.BodyMeasurement.id == id).first()
     if db_measurement is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f'body measurement with id {id} not found')
+    if db_measurement.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail='Not authorized to perform the requested action')
     db.delete(db_measurement)
     db.commit()
     return {'message':'successfully deleted the body measurement'}
